Start working on example which reads samples and filter them, there is one workaround, because conversion is done by reading 4B and typecast them to float, this is supposed to run on PC side in simulation and on arduino, to be possible this I created auxiliary VI for retyping byte array to float (there is explanation in it's description).
To convert 4B array to float I need:
  - on arduino convert it to Variant and after to float
  - on PC type cast it
This is different, so I made VI which transcript code for arduino using retyping to variant and after to float, inside in LabVIEW is implemented as typecast so what is written above is its exact behaviour.
This couldn't not be resolved probably other way. (but still this is quite simple, when I tried on PC side convert to variant and after to float it's not giving right results).